restructure the location of the render tool to cmd/render and the others to cmd/controller
