+++ cmd/render/.goreleaser.yaml
